chore: rename settings page
