+++ app/src/views/features/sessions/SessionsSettingsPage/PageSourceRow/pageSourceRow.css
